+++ backend/zog_igame/__init__.py
@@ -2,5 +2,5 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import models
-from . import controllers
+# from . import controllers
