+++ backend/zog_igame/models/board.py
@@ -20,12 +20,10 @@ class Board(models.Model):
     _name = "og.board"
     _description = "Board"
     _rec_name = 'number'
-    _order = 'number'   #_order: 定义search()和read()方法的结果记录的排序规则，
-    # 和SQL语句中的order 类似，缺省值是id,即按id升序排序
+    _order = 'number'
 
     table_id = fields.Many2one('og.table')
     deal_id = fields.Many2one('og.deal')
-
 
     name   = fields.Char('Name', related='deal_id.name' )
     number = fields.Integer(related='deal_id.number' )
