+++ backend/zog_igame/views/deal_views.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <odoo>
         <menuitem name="Deal"
             id="menu_deal"
