+++ backend/zog_igame/views/igame_views.xml
@@ -34,50 +34,61 @@
                     <field name="match_type"/>
                     <field name="org_type"/>
                     <field name="score_uom"/>
-
                 </tree>
             </field>
         </record>
 
-        <record id="view_og_igame_score_tree" model="ir.ui.view">
-            <field name="name">igame.score.tree</field>
-            <field name="model">og.igame.score</field>
+        <record id="view_og_igame_team_tree" model="ir.ui.view">
+            <field name="name">igame.team.tree</field>
+            <field name="model">og.igame.team</field>
             <field name="arch" type="xml">
-                <tree string="Intelligent Game Score">
+                <tree string="Intelligent Game Team">
                     <field name="igame_id"/>
                     <field name="partner_id"/>
-                    <field name="number"/>
-                    <field name="position"/>
-                    <field name="score"/>
-                    <field name="score_uom"/>
-                    <field name="rank"/>
-
-                </tree>
-            </field>
-        </record>
-
-        <record id="view_og_igame_score_line_tree" model="ir.ui.view">
-            <field name="name">igame.score.line.tree</field>
-            <field name="model">og.igame.score.line</field>
-            <field name="arch" type="xml">
-                <tree string="Intelligent Game Score Line">
-                    <field name="igame_id"/>
                     <field name="group_id"/>
-                    <field name="partner_id"/>
-                    <field name="round_id"/>
-                    <field name="position"/>
-                    <field name="deal_id"/>
+                    <field name="number"/>
+                    <!--<field name="position"/>-->
                     <field name="score"/>
                     <field name="score_uom"/>
-                    <field name="rank"/>
+
+                </tree>
+            </field>
+        </record>
+
+        <record id="view_og_igame_team_player_tree" model="ir.ui.view">
+            <field name="name">igame.team.player.tree</field>
+            <field name="model">og.igame.team.player</field>
+            <field name="arch" type="xml">
+                <tree string="Intelligent Game Team Player">
+                    <field name="partner_id"/>
+                    <field name="role"/>
+                    <field name="team_id"/>
+
+
+                </tree>
+            </field>
+        </record>
+
+        <record id="view_og_igame_team_line_tree" model="ir.ui.view">
+            <field name="name">igame.team.line.tree</field>
+            <field name="model">og.igame.team.line</field>
+            <field name="arch" type="xml">
+                <tree string="Intelligent Game Team Line">
+                    <field name="igame_id"/>
+
+                    <field name="team_id"/>
+                    <field name="round_id"/>
+
+                    <field name="score"/>
+                    <field name="score_uom"/>
 
                 </tree>
             </field>
         </record>
 
         <record model="ir.actions.act_window" id="action_igame_score_list">
-            <field name="name">Ientelligent Game Score</field>
-            <field name="res_model">og.igame.score</field>
+            <field name="name">Ientelligent Game Team</field>
+            <field name="res_model">og.igame.team</field>
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
             <field name="help" type="html">
@@ -87,7 +98,7 @@
         </record>
 
         <menuitem id="menu_game_score_list"
-            name="iGames Score"
+            name="iGames Team"
             parent="menu_igame"
             action="action_igame_score_list"
             sequence="30"/>
